textfield, tooltip, header, helpdrawer migrated to sb7 format
